# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sunrise Staff Housing Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assignments/in-house:
    get:
      operationId: listInHouseAssignments
      tags: [assignments]
      summary: List in-house assignments (paginated)
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: buildingId
          in: query
          schema:
            type: integer
        - name: floorId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Paginated list of assignments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Assignment"
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: roomId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create assignment (check-in employee)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentBody"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    get:
      operationId: getAssignment
      tags: [assignments]
      summary: Get assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    patch:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssignmentBody"
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}/checkout:
    post:
      operationId: checkoutAssignment
      tags: [assignments]
      summary: Check out employee from room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutBody"
      responses:
        "200":
          description: Checked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}/transfer:
    post:
      operationId: transferAssignment
      tags: [assignments]
      summary: Transfer employee to another room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferBody"
      responses:
        "200":
          description: Transferred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"

  # RESERVATIONS
...
components:
  schemas:
...
    Assignment:
      type: object
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        employeeId:
          type: integer
        roomId:
          type: integer
        bedNumber:
          type: ["integer", "null"]
        checkInDate:
          type: string
        expectedCheckOutDate:
          type: ["string", "null"]
        checkOutDate:
          type: ["string", "null"]
        notes:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - propertyId
        - employeeId
        - roomId
        - checkInDate
        - notes
        - status
        - createdAt
    CreateAssignmentBody:
      type: object
      properties:
        propertyId:
          type: integer
        employeeId:
          type: integer
        roomId:
          type: integer
        bedNumber:
          type: ["integer", "null"]
        checkInDate:
          type: string
        expectedCheckOutDate:
          type: ["string", "null"]
        notes:
          type: string
      required:
        - propertyId
        - employeeId
        - roomId
        - checkInDate
    UpdateAssignmentBody:
      type: object
      properties:
        expectedCheckOutDate:
          type: ["string", "null"]
        notes:
          type: string
    CheckoutBody:
      type: object
      properties:
        checkOutDate:
          type: string
        notes:
          type: string
      required:
        - checkOutDate
    TransferBody:
      type: object
      properties:
        newRoomId:
          type: integer
        newBedNumber:
          type: integer
        transferReason:
          type: string
        transferDate:
          type: string
      required:
        - newRoomId
        - transferDate
...